requirements.txt
torch
torchvision
numpy
faiss-cpu
pillow
tqdm
icrawler